Convert fetching site orders status to suspendable call
